fix: 동물 선택 기능 구현 수정 및 컴포넌트 분리
